move: Move old utils to lib
